mon/OSDMonitor: fix 'osd pool set ...' int/float parsing
In 2fe0d0d97af95c22db80800f5b9da51f672d9407 I assumed that the command
may come with a range of JSON formats (string, int, float), but in
reality the client gets the schema via get_command_descriptions
regardless of what version it is running.  That means we only need to
handle the string case.  This fixes warnings like

2013-10-29 09:04:23.911996 7f38df7fe700 -1 bad boost::get: key val is not type long
2013-10-29 09:04:23.934971 7f38df7fe700 -1 0x7f38df7f9578
2013-10-29 09:04:23.935014 7f38df7fe700 -1 bad boost::get: key val is not type float
2013-10-29 09:04:23.946177 7f38df7fe700 -1 0x7f38df7f9578

which are noted in #6673.

Signed-off-by: Sage Weil <sage@inktank.com>