install.upgrade: fix overrides of sha1|tag|branch
If the upgrade task config has a branch: (or tag or sha1), do
not apply the sha1|branch|tag overrides keys.  This fixes the
breakage from 280f783c2e8dda0df6afb4de0b115aad1614fbdc which
made

overrides:
  install:
    ceph:
      sha1: f88866fe9c218e03b3f89ff21e2b7b655aa74294
tasks:
...
- install.upgrade:
    all:
      branch: dumpling

...use the sha1 from the overrides instead of the explicitly
specified branch.  The intention was to only use the overrides
when the version was not specified (whether it was sha1 or
branch or tag).

At some point we should probably make the same change for
install function in install.py, but let's fix this first to
get the upgrade tests working.

Signed-off-by: Sage Weil <sage@inktank.com>