BSD-Linux interoperability
Ceph's networking protocol passes the in-memory representation of struct
sockaddr_storage over the wire, even though that structure is
operating-system and architecture specific.  This commit creates a
portably-defined struct ceph_sockaddr_storage which is identical to struct
sockaddr_storage on Linux amd64, along with appropriate encode and decode
functions.  This allows FreeBSD and Linux machines to coexist on the same
Ceph cluster.

Signed-off-by: Alan Somers <asomers@gmail.com>