mds: include inode version in auth mds's lock messages
encode inode version in auth mds's lock messages, so that version
of replica inodes get updated. This is important because client
use inode version in mds reply to check if the cached inode is
already up-to-date. It skips updating the cached inode if it thinks
the cached inode is up-to-date.

Signed-off-by: Yan, Zheng <zheng.z.yan@intel.com>