erasure-code: jerasure implementation of chunk size helpers
The encode methode uses the get_chunk_size method which is a translation
of the code it previously used for the same purpose. The chunk_count and
chunk_data_count methods return k+m and k respectively. But that will
no longer be the case after local parity is implemented.

Signed-off-by: Loic Dachary <loic@dachary.org>