OSDMap: add primary out params to internal _pg_to_up_acting_osds function
And use pointers instead of references for out params.

Now pg_to_up_acting_osds and pg_to_acting_osds can plug in to this slightly
more real implementation, instead of making up their own. (We are still
just using the first member anyway, but we're about to plug it into
the bottom layer of functions.)

Signed-off-by: Greg Farnum <greg@inktank.com>