SimpleMessenger: do not hold lock for submit_message()
We were taking the messenger lock for the duration of submit_message(), but
only actually need it if we make use of our global maps -- often we complete
the function just via the Connection and Pipe we pass in. So only take
the lock if we need it.

TODO: verify that this works properly under reconnect situations, etc. I'm
not actually sure that we're race-free.