OSDMonitor: make sure we don't send out maps with a primary_temp mapping
Making sure a cluster supports primary_temp is complicated and we don't
have any of the machinery in place right now (nor a need to actually support
it). We don't have any mechanisms for setting it to begin with, so assert
that we never create anything with any such mapping in update_from_paxos()
to catch any errors.

Signed-off-by: Greg Farnum <greg@inktank.com>