osd: ignore num_objects_dirty for old pools
Way back in b191d1e05a8adb992230874fb770cb15b5a1b93a we introduced the
dirty flag, but we did not track it in the stats until much later in
1f4752d7530d954800e13c09254796ed7210ed4b.  Unfortunately this interval
spans the emperor release.  To avoid making scrub error out and require
repair on *any* of those old pools, flag stats that were encoded before
now such that the dirty stats are ignored.  Clear the flag if we *do*
do a repair so that it will be tracked properly thereafter.

Signed-off-by: Sage Weil <sage@inktank.com>