FileStore: The getattr and read interface changed
getattr and read ObjectStore interfaces are changed to accomodate two
new parameters, the file descriptor and the full filesystem path.
For getattr, these two out params are default to NULL and if the caller
is populating those other than NULL value, getattr will send the caller
the file descriptor and the file system path.
For read, these two in params are default to -1 and NULL respectively.
If caller supplies valid values for fd , it will try to read from
that fd first. Upon failing, it will try to open directly the filepath
(if not empty). If that also fails, it is calling lfn_open().
There is a config option filestore_use_fd_cache to switch between
old and new interface of lfn_open().

Signed-off-by: Somnath Roy <somnath.roy@sandisk.com>