erasure-code: new interface methods encode/decode_chunks
The encode and decode interface are expected to allocate properly
aligned chunks when needed and convert bufferlists into chunks. The
encode_chunks and decode_chunks are lower level interfaces that assume
all chunks are allocated and that all chunks are to be decoded or
encoded.

They are meant to be used in contexts where these constraints are
enforced by default, such as when a plugin is used within the pyramid
erasure code implementation.

Reviewed-By: Christophe Courtaut <christophe.courtaut@gmail.com>
Signed-off-by: Loic Dachary <loic@dachary.org>